# Reuse root context in a sync pool to reduce GC: zero-alloc routing