… _examples/rest/render/render.go