@pkieltyka 2016-03-30
btw.. for fun I ran the benchmarks on Go-tip (1.7)

```shell
BenchmarkChi_Param                	10000000	       141 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_Param               	 2000000	       671 ns/op	     336 B/op	       2 allocs/op
BenchmarkChi_Param5               	 5000000	       380 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_Param5              	 2000000	      1004 ns/op	     336 B/op	       2 allocs/op
BenchmarkChi_Param20              	 1000000	      1301 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_Param20             	  500000	      2985 ns/op	    1247 B/op	       2 allocs/op
BenchmarkChi_ParamWrite           	10000000	       190 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_ParamWrite          	 2000000	       743 ns/op	     336 B/op	       2 allocs/op
BenchmarkChi_GithubStatic         	10000000	       219 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_GithubStatic        	10000000	       238 ns/op	       0 B/op	       0 allocs/op
BenchmarkChi_GithubParam          	 3000000	       445 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_GithubParam         	 1000000	      1074 ns/op	     336 B/op	       2 allocs/op
BenchmarkChi_GithubAll            	   20000	     79445 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_GithubAll           	    3000	    466811 ns/op	   56116 B/op	     334 allocs/op
BenchmarkChi_GPlusStatic          	10000000	       139 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_GPlusStatic         	10000000	       168 ns/op	       0 B/op	       0 allocs/op
BenchmarkChi_GPlusParam           	10000000	       194 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_GPlusParam          	 2000000	       728 ns/op	     336 B/op	       2 allocs/op
BenchmarkChi_GPlus2Params         	 5000000	       283 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_GPlus2Params        	 1000000	      1030 ns/op	     336 B/op	       2 allocs/op
BenchmarkChi_GPlusAll             	  500000	      3047 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_GPlusAll            	  200000	     10031 ns/op	    3696 B/op	      22 allocs/op
BenchmarkChi_ParseStatic          	10000000	       140 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_ParseStatic         	10000000	       215 ns/op	       0 B/op	       0 allocs/op
BenchmarkChi_ParseParam           	10000000	       175 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_ParseParam          	 2000000	       796 ns/op	     336 B/op	       2 allocs/op
BenchmarkChi_Parse2Params         	10000000	       230 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_Parse2Params        	 2000000	       800 ns/op	     336 B/op	       2 allocs/op
BenchmarkChi_ParseAll             	  300000	      5261 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_ParseAll            	  100000	     16631 ns/op	    5376 B/op	      32 allocs/op
BenchmarkChi_StaticAll            	   30000	     47631 ns/op	       0 B/op	       0 allocs/op
BenchmarkGoji_StaticAll           	   30000	     52332 ns/op	       0 B/op	       0 allocs/op
```